:sparkles: guest라우터
ios를 위해서 만들어봤습니다.
