+++ src/Club/club.service.ts
@@ -310,72 +310,6 @@ export class ClubService {
     });
   }
 
-  async guestDetailPage(clubType: string, clubName: string) {
-    const club = await this.Club.findOne({
-      where: { type: clubType, title: clubName },
-      relations: ['activityUrls', 'relatedLink', 'member', 'member.user'],
-      select: {
-        id: true,
-        title: true,
-        type: true,
-        bannerUrl: true,
-        description: true,
-        contact: true,
-        teacher: true,
-        isOpened: true,
-        member: {
-          id: true,
-          user: {
-            email: true,
-            name: true,
-            grade: true,
-            class: true,
-            num: true,
-            userImg: true,
-          },
-          scope: true,
-        },
-        relatedLink: { name: true, url: true, id: true },
-        activityUrls: { id: true, url: true },
-      },
-    });
-
-    if (!club)
-      throw new HttpException(
-        '존재하지 않는 동아리입니다.',
-        HttpStatus.NOT_FOUND,
-      );
-    const head = club.member.find((member) => {
-      return member.scope === 'HEAD';
-    });
-
-    const clubMembers = club.member.filter((member) => {
-      return member.scope === 'MEMBER';
-    });
-
-    if (club.activityUrls) {
-      const activityurls = club.activityUrls.map((url) => {
-        return url.url;
-      });
-      const isApplied = false;
-      const scope = 'USER';
-
-      delete club.member;
-      delete club.activityUrls;
-
-      return {
-        club,
-        activityurls,
-        head: head.user,
-        member: clubMembers.map((user) => {
-          return user.user;
-        }),
-        scope,
-        isApplied,
-      };
-    }
-  }
-
   async detailPage(clubtype: string, clubtitle: string, email: string) {
     const club = await this.Club.findOne({
       where: { type: clubtype, title: clubtitle },
